week1/requirements.txt
langchain==0.1.0
langchain-community>=0.0.10
flask==2.3.3
python-dotenv==1.0.0
fastapi
requests>=2.31.0
uvicorn
python-multipart
uuid
openai>=1.0.0
